teedogs first Instant-Auto-Merge from BRANCH-1-9 to HEAD: please dont break
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk/tiki@5265 b456876b-0849-0410-b77d-98878d47e9d5

+++ tiki-admin_system.php
@@ -1,5 +1,5 @@
 <?php
-// $Header: /cvsroot/tikiwiki/tiki/tiki-admin_system.php,v 1.15 2004-05-06 00:47:10 mose Exp $
+// $Header: /cvsroot/tikiwiki/tiki/tiki-admin_system.php,v 1.16 2004-08-12 22:31:21 teedog Exp $
 
 // Copyright (c) 2002-2004, Luis Argerich, Garland Foster, Eduardo Polidor, et. al.
 // All Rights Reserved. See copyright.txt for details and a complete list of authors.
@@ -39,7 +39,7 @@ function erase_dir_content($path) {
 	if (!$path or !is_dir($path)) return 0;
 	if ($dir = opendir($path)) {
 		while (false !== ($file = readdir($dir))) {
-			if (substr($file,0,1) == "." or $file == 'CVS') continue;
+			if (substr($file,0,1) == "." or $file == 'CVS' or $file == "index.php" or $file == "README") continue;
 			if (is_dir($path."/".$file)) {
 				erase_dir_content($path."/".$file);
 				rmdir($path."/".$file);
@@ -132,7 +132,7 @@ foreach($languages as $clang) {
 	if(is_dir("templates_c/$tikidomain/".$clang["value"])) {
 		$templates[$clang["value"]] = du("templates_c/$tikidomain/".$clang["value"]);
 	} else {
-		$templates[$clang["value"]] = du("templates_c/", substr($tikidomain.$clang["value"], 1));
+		$templates[$clang["value"]] = du("templates_c/", $tikidomain.$clang["value"]);
 	}
 }
 $smarty->assign_by_ref('templates', $templates);
